cultural-fest/backend/requirements.txt
fastapi==0.136.0
uvicorn[standard]==0.44.0
supabase==2.28.3
python-dotenv==1.2.2
Pillow==12.2.0
qrcode==8.2
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
python-multipart==0.0.26
pydantic==2.13.3
